Add next plan item
In addition to the previous plan item addition about configuring error
checking, we're adding a new plan item for handling command line
parameters.

Also updated the current step number to indicate that we've completed
the error checking configuration. That configuration is entirely outside
of the codebase, however, being provided as option parameters to CMake
when creating the build files.